requirements.txt
setuptools>=31.0.1
Babel==2.5.0
Django==1.9.4
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
PyJWT==1.5.2
PyYAML==3.12
Pygments==2.2.0
South==1.0.2
Sphinx==1.6.3
alabaster==0.7.10
amqp==1.4.9
anyjson==0.3.3
argh==0.26.2
argparse==1.4.0
asn1crypto==0.22.0
autopep8==1.3.2
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
billiard==3.3.0.23
bottlenose==1.1.3
bs4==0.0.1
caniusepython3==5.0.0
celery==3.1.25
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.1
coreschema==0.0.4
coverage==4.4.1
cryptography==1.9
defusedxml==0.5.0
distlib==0.2.5
django-cachalot==1.5.0
django-celery==3.2.1
django-celery-beat==1.0.1
django-extensions==1.8.1
django-picklefield==1.0.0
django-rest-swagger==2.1.2
django-webpack-loader==0.5.0
djangorestframework==3.6.3
docutils==0.14
ecdsa==0.13
fake-useragent==0.1.7
flake8==3.4.1
future==0.16.0
futures==3.1.1
fuzzywuzzy==0.15.1
google-api-python-client==1.6.2
gunicorn==19.7.1
hashids==1.2.0
html5lib==0.999999999
http-ece==1.0.4
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
interruptingcow==0.7
itypes==1.1.0
jsondiff==1.1.1
jsonfield==2.0.2
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
kombu==3.0.37
livereload==2.5.1
lxml==3.8.0
mccabe==0.6.1
mock==2.0.0
nltk==3.2.4
numpy==1.13.1
oauth==1.0.1
oauth2client==4.1.2
oauthlib==2.0.2
openapi-codec==1.3.2
packaging==16.8
pathtools==0.1.2
pbr==3.1.1
pep8==1.7.0
port-for==0.3.1
prettytable==0.7.2
progressbar2==3.34.2
psycopg2>=2.8,<2.9
py-vapid==1.2.5
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
pyelliptic==1.5.8
pyflakes==1.6.0
pyparsing==2.2.0
pyserial==3.4
python-amazon-simple-product-api==2.2.11
python-dateutil==2.6.1
python-jose==1.3.2
python-memcached==1.58
python-openid==2.2.5
python-social-auth==0.2.21
python-utils==2.2.0
python3-openid==3.1.0
pytz==2017.2
pywebpush==1.1.0
recommonmark==0.4.0
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rollbar==0.16.1
rsa==3.4.2
scikit-learn==0.19.0
scipy==0.19.1
selenium==3.5.0
simplejson==3.11.1
six==1.10.0
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.2.4
sphinxcontrib-inlinesyntaxhighlight==0.2
sphinxcontrib-websupport==1.0.1
ssh-import-id==5.6
toolz==0.8.2
tornado==4.5.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
watchdog==0.8.3
webencodings==0.5.1
